# Read file in do_hash_file as Bytes and optimize read_file size_hint on Windows