Reading the file in do_hash_file as Bytes skips CRLF to LF conversion and is therefore way faster.
Optimize read_file size_hint on Windows by getting the file size of the existing handle instead of opening a duplicate one.