@clemenswasser 2024-03-20
Some performance numbers for a completely cached CMake clean rebuild (includes linking time, so impact is bigger than it seems)
before:
```console
> hyperfine -p 'cmake --build .\build-msvc-debug\ -t clean' 'cmake --build .\build-msvc-debug\'
Benchmark 1: cmake --build .\build-msvc-debug\
  Time (mean ± σ):      7.530 s ±  0.110 s    [User: 14.683 s, System: 10.512 s]
  Range (min … max):    7.304 s …  7.690 s    10 runs
```
after:
```console
> hyperfine -p 'cmake --build .\build-msvc-debug\ -t clean' 'cmake --build .\build-msvc-debug\'
Benchmark 1: cmake --build .\build-msvc-debug\
  Time (mean ± σ):      7.232 s ±  0.099 s    [User: 10.662 s, System: 7.441 s]
  Range (min … max):    7.096 s …  7.390 s    10 runs
```