@clemenswasser 2024-03-20
I've used the `.ninja_log` file to look how much this saves without linking, etc., by summing all durations of the obj files.
Before: 61737ms
After: 53393ms
So for a completely cached CMake rebuild, ccache is now ~1.16 times faster on Windows :^)